requirements.txt
torch>=1.9.0
transformers>=4.20.0
numpy>=1.21.0
nltk>=3.7
sentence-transformers>=2.2.0
spacy>=3.4.0
textstat>=0.7.3
google-generativeai>=0.3.0
language-tool-python>=2.7.1
gradio>=4.0.0
plotly>=5.15.0
scipy>=1.7.0
python-dotenv>=1.0.0